backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
sqlalchemy>=2.0
pydantic>=2.7
pandas>=2.2
openpyxl>=3.1
python-dotenv>=1.0
alembic>=1.13
pymysql>=1.1.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7
